Translates global positioning data to local accelerations
Takes ANSYS global motion data and translates into a local reference frame, as if there was an MRU at a position on the body. The MRU can be in any direction, but rot_z is likely the only rotation that needs to be changed. The motion of the accelerometer can be plotted with .animate_plot(). The acceleration, rotation and rotational accelerations can be saved to an excel file with .save_data("FILE NAME").